% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{calendar}
\format{
A data.table with the following columns:
\describe{
\item{\code{ dte }}{ the date in \emph{standard} format \code{YYYY-MM-DD} }
\item{\code{ is_weekday }}{  }
\item{\code{ is_workingday }}{  }
\item{\code{ year }}{  }
\item{\code{ fyear }}{   }
\item{\code{ quarter }}{   }
\item{\code{ qoy }}{  }
\item{\code{ month }}{  }
\item{\code{ moy }}{  }
\item{\code{ moyt }}{  }
\item{\code{ moyts }}{  }
\item{\code{ moq }}{  }
\item{\code{ week }}{  }
\item{\code{ dweek }}{  }
\item{\code{ woy }}{  }
\item{\code{ woq }}{  }
\item{\code{ wom }}{  }
\item{\code{ doy }}{  }
\item{\code{ doq }}{  }
\item{\code{ dom }}{  }
\item{\code{ dow }}{ Day of the Week (following ISO-8601 with Monday the first day)  }
\item{\code{ dowt }}{ Day of the Week as text, an ordered factor with Monday as the first day }
\item{\code{ dowts  }}{ Day of the Week as short text, an ordered factor with Monday as the first day }
\item{\code{ rdow }}{ Day of the Week (following most religions with Sunday the first day) }
\item{\code{ rdowt }}{ Day of the Week as text, an ordered factor with Sunday as the first day }
\item{\code{ rdowts  }}{ Day of the Week as short text, an ordered factor with Sunday as the first day }
\item{\code{ mseason }}{ \emph{Meteorological} Season: the starts are the first of the months: March, June, September, December }
\item{\code{ aseason }}{ \emph{Astronomical} Season: the starts are: 20th March, 21th June, 23th September, 21th December }
}
}
\usage{
calendar
}
\description{
Description ...
}
\keyword{datasets}
