% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcalendar.R
\docType{package}
\name{Rcalendar}
\alias{Rcalendar}
\title{Rcalendar}
\description{
A simple package containing a lookup table \code{calendar} between different ways of expressing dates.
It also includes a list of the formulas used for each representation
}
