% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{Seasons}
\format{
A data.table with the following columns:
\describe{
\item{\code{ year }}{ the year }
\item{\code{ spring }}{ the starting day for Spring }
\item{\code{ summer }}{ the starting day for Summer }
\item{\code{ autumn }}{ the starting day for Autumn }
\item{\code{ winter }}{ the starting day for Winter }
}
}
\usage{
seasons
}
\description{
Starting Day for the seasons of the year, from 1970 to 2050.
}
\details{
Based on the \href{https://neoprogrammics.com/equinoxes_and_solstices/Equinoxes_and_Solstices_Calculator.php}{600-Year NASA/JPL DE405 Ephemeris Model}
}
\keyword{datasets}
