% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_holidays}
\alias{uk_holidays}
\title{UK holidays}
\format{
A data.table with the following columns:
\describe{
\item{\code{ dte }}{ the date }
\item{\code{ name }}{  the name of the holiday }
}
}
\usage{
uk_holidays
}
\description{
A list of all UK Bank Holidays from 1970 to 2025
}
\keyword{datasets}
