% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{calendar}
\format{
A data.table with the following columns:
\describe{
\item{\code{ dte }}{ the date in \emph{standard} format \code{YYYY-MM-DD} }
\item{\code{ is_weekday }}{ \code{Yes} for Mon-Fri, \code{No} for Sat and Sun }
\item{\code{ is_workingday }}{ \code{No} for Sat and Sun plus \emph{holidays} (as listed in the \code{holidays_uk} table), \code{Yes} for all the other days }
\item{\code{ year }}{ Year }
\item{\code{ fyear }}{ Financial Year (1st April to 31st March)  }
\item{\code{ quarter }}{ Quarter (as Year + Quarter of the Year) }
\item{\code{ qoy }}{ Quarter of the Year }
\item{\code{ month }}{ Month (as Month of the Year in short form text + Year) }
\item{\code{ moy }}{ Month of the Year }
\item{\code{ moyt }}{ Month of the Year in text  }
\item{\code{ moyts }}{ Month of the Year in short form text  }
\item{\code{ moq }}{ Month of the Quarter }
\item{\code{ week }}{ Week of the Year (as Year + \code{-W} + the number of the Week following ISO 8601) }
\item{\code{ dweek }}{ Date of the Week of the Year (as the standard date of the Monday of the Week of the Year) }
\item{\code{ woy }}{ Week of the Year (number only) }
\item{\code{ woq }}{ Week of the Quarter }
\item{\code{ wom }}{ Week of the Month }
\item{\code{ doy }}{ Day of the Year }
\item{\code{ doq }}{ Day of the Quarter }
\item{\code{ dom }}{ Day of the Month }
\item{\code{ dow }}{ Day of the Week (following ISO-8601, with Monday the first day of the Week)  }
\item{\code{ dowt }}{ Day of the Week as text, an ordered factor with Monday as the first day of the Week }
\item{\code{ dowts  }}{ Day of the Week as short text, an ordered factor with Monday as the first day of the Week }
\item{\code{ gdow }}{ Gregorian Day of the Week (with Sunday the first day of the Week) }
\item{\code{ gdowt }}{ Gregorian Day of the Week as text, an ordered factor with Sunday as the first day of the Week }
\item{\code{ gdowts  }}{ Gregorian Day of the Week as short text, an ordered factor with Sunday as the first day of the Week }
\item{\code{ season }}{ \emph{Astronomical} Season (as listed in the \code{seasons} table) }
\item{\code{ mseason }}{ \emph{Meteorological} Season (with the starting days being the first of the months: March, June, September, December) }
}
}
\usage{
calendar
}
\description{
All days from 1st Jan 1970 until 31st Dec 2025 with some properties, characteristics and formattings.
}
\keyword{datasets}
